feat: graphs fixed

+++ sem4/graph/common/src/Generator.tpp
@@ -44,7 +44,7 @@ std::unique_ptr<GraphT> Generator::generateAcyclicTemplate(int numVertices, int 
     std::shuffle(candidates.begin(), candidates.end(), rng_);
 
     int to_add = numEdges - (numVertices - 1);
-    for (int i = 0; i < to_add && i < candidates.size()); ++i) {
+    for (int i = 0; i < to_add && i < static_cast<int>(candidates.size()); ++i) {
         auto [from, to] = candidates[i];
         addEdge(*graph, from, to, genWeight());
     }
@@ -61,7 +61,7 @@ std::unique_ptr<GraphT> Generator::generateByDegreesTemplate(int numVertices,
 {
     if (numVertices <= 0)
         throw std::invalid_argument("Количество вершин должно быть положительным");
-    if (degrees.size() != numVertices)
+    if (degrees.size() != static_cast<std::size_t>(numVertices))
         throw std::invalid_argument("Размер вектора степеней должен совпадать с числом вершин");
 
     auto graph = std::make_unique<GraphT>(isDirected);
@@ -74,7 +74,7 @@ std::unique_ptr<GraphT> Generator::generateByDegreesTemplate(int numVertices,
 
         int from;
         if (!with_budget.empty()) {
-            from = with_budget[randomInt(0, with_budget.size() - 1)];
+            from = with_budget[randomInt(0, static_cast<int>(with_budget.size()) - 1)];
         } else {
             from = randomInt(0, i - 1);
         }
@@ -90,7 +90,7 @@ std::unique_ptr<GraphT> Generator::generateByDegreesTemplate(int numVertices,
             if (!graph->hasEdge(i, j)) available.push_back(j);
 
         std::shuffle(available.begin(), available.end(), rng_);
-        int toAdd = std::min(degrees[i], available.size());
+        int toAdd = std::min(degrees[i], static_cast<int>(available.size()));
 
         for (int k = 0; k < toAdd; ++k)
             addEdge(*graph, i, available[k], genWeight());
